Refactor to start Android support
